Update to Swift 1.1 (Xcode 6.1)

+++ pulse-swiftTests/bep/v1/CloseTests.swift
@@ -6,7 +6,7 @@ class CloseTests : XCTestCase {
     let close = Close(reason: someReason)
 
     func testHasType7() {
-        XCTAssertEqual(close.type, 7)
+        XCTAssertEqual(close.type, UInt8(7))
     }
 
     func testHasXdrEncodedReason() {
