+++ pulse-swiftTests/bep/v1/IndexUpdateTests.swift
@@ -4,7 +4,7 @@ import pulse
 class IndexUpdateTests : XCTestCase {
 
     func testIsType6() {
-        XCTAssertEqual(IndexUpdate(folder: "", files: []).type, 6)
+        XCTAssertEqual(IndexUpdate(folder: "", files: []).type, UInt8(6))
     }
 
 }
