+++ pulse-swiftTests/bep/v1/PingTests.swift
@@ -6,7 +6,7 @@ class PingTests : XCTestCase {
     let ping = Ping()
 
     func testHasType4() {
-        XCTAssertEqual(ping.type, 4)
+        XCTAssertEqual(ping.type, UInt8(4))
     }
 
     func testIsNotCompressed() {
